rename folders to _hw
