+++ Lessons/lecture_19_hw/assert_calc.py
@@ -2,7 +2,7 @@ import unittest
 from parameterized import parameterized
 
 
-from ..lecture_18.app import calc
+from ..lecture_18_hw.app import calc
 
 # _calc = calc.Calculator()
